Remove Nachlauf max limit; support >255 lines via chunked ESC d
- Removed max="20" from prPostFeed input (no limit)
- postFeedLines changed to uint16_t throughout (PrinterManager, LabelRenderer)
- LabelRenderer: ESC d chunked in ≤255-line steps to handle large values

+++ src/printer/LabelRenderer.h
@@ -7,12 +7,12 @@
 class LabelRenderer {
 public:
     explicit LabelRenderer(EscPosPrinter &printer);
-    void setPreFeed(uint8_t lines)  { preFeedLines  = lines; }
-    void setPostFeed(uint8_t lines) { postFeedLines = lines; }
+    void setPreFeed(uint8_t lines)   { preFeedLines  = lines; }
+    void setPostFeed(uint16_t lines) { postFeedLines = lines; }
     bool printInventoryLabel(const InventoryItem &item);
 
 private:
     EscPosPrinter &printer;
-    uint8_t preFeedLines  = LABEL_PRE_FEED;
-    uint8_t postFeedLines = 4;
+    uint8_t  preFeedLines  = LABEL_PRE_FEED;
+    uint16_t postFeedLines = 4;
 };
