+++ src/printer/PrinterManager.h
@@ -14,7 +14,7 @@ public:
     void feed(uint8_t lines);
     void backFeed(uint8_t lines);
     void setBackfeedConfig(bool enabled, uint16_t dots);
-    void setPostFeed(uint8_t lines);
+    void setPostFeed(uint16_t lines);
     // Raw test: cmd=0x6A/0x4B/0x65, dots, chunkSize(0=single), delayMs, flush
     bool testBackfeed(uint8_t cmd, uint16_t dots, uint8_t chunkSize, uint16_t delayMs, bool doFlush);
     bool printLabel(const InventoryItem &item);
@@ -28,5 +28,5 @@ private:
     uint32_t currentBaud     = 0;
     bool     backfeedEnabled = false;
     uint16_t backfeedDots    = 72;
-    uint8_t  postFeedLines   = 4;
+    uint16_t postFeedLines   = 4;
 };
